When canceling the log dialog: if there's a timeout on the first cancel try just return. Only if the timeout persists on the second try exit the process. Reason: setting the cancel flag and waiting for the threads to exit might never happen since at least the log thread accesses the UI and therefore needs the UI thread to handle the message loop. But when waiting for the threads to exit, the message loop isn't handled so there's a deadlock situation.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@25403 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlg.cpp
@@ -1218,11 +1218,6 @@ void CLogDlg::SaveSplitterPos()
 
 void CLogDlg::OnLogCancel()
 {
-    // canceling means stopping the working thread if it's still running.
-    // we do this by using the Subversion cancel callback.
-
-    m_bCancelled = true;
-
     // Canceling can mean just to stop fetching data, depending on the text
     // shown on the cancel button (it will read "Cancel" in that case).
 
@@ -1230,7 +1225,10 @@ void CLogDlg::OnLogCancel()
     GetDlgItemText(IDC_LOGCANCEL, temp);
     temp2.LoadString(IDS_MSGBOX_CANCEL);
     if ((temp.Compare(temp2)==0) && !GetDlgItem(IDOK)->IsWindowVisible())
+    {
+        m_bCancelled = true;
         return;
+    }
 
     // we actually want to close the dialog.
 
@@ -1239,6 +1237,7 @@ void CLogDlg::OnLogCancel()
 
 void CLogDlg::OnCancel()
 {
+    bool bWasCancelled = m_bCancelled;
     // canceling means stopping the working thread if it's still running.
     // we do this by using the Subversion cancel callback.
 
@@ -1255,8 +1254,16 @@ void CLogDlg::OnCancel()
 
     if (threadsStillRunning)
     {
-        // end the process the hard way
-        TerminateProcess(GetCurrentProcess(), 0);
+        if (bWasCancelled)
+        {
+            // end the process the hard way
+            TerminateProcess(GetCurrentProcess(), 0);
+        }
+        else
+        {
+            // to force the shutdown, another click on cancel is required
+            return;
+        }
     }
     else
     {
